First conventions
Here I'm proposing two conventions, one for the vocabulary and the other for the style. The first one is to use the word changeset as a noun which, for non-native speakers, simplifies the (mental) parsing a little bit. The second is to use they/their/them for the neutral gender instead of she/her/her *regardless* of the number (singular/plural). Even though both forms are widely used I've seen more usage of the third person in recent times. Please notice that these are nothing but proposals and it is up to you whether to adopt them or not.